server/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-api-python-client
cloud-sql-python-connector
SQLAlchemy
python-tds
sqlalchemy-pytds
requests
json